# services/onboarding-service/specs/apiSpecs.yaml
openapi: 3.0.3
info:
  title: IBAN Validation API
  version: 1.0.0
  description: API to validate an IBAN and retrieve bank information.
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api-dev.alramz.com/validation
    description: Development Environment
  - url: https://api-test.alramz.com/validation
    description: Test Environment
  - url: https://api-preprod.alramz.com/validation
    description: Pre-Production Environment
  - url: https://api.alramz.com/validation
    description: Production Environment
security:
  - BearerAuth: []
paths:
  /iban/validate:
    post:
      tags:
        - IBAN
      summary: Validate IBAN
      description: Validates an IBAN and returns bank details along with validation results.
      operationId: validateIBAN

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IBANRequest'
            example:
              IBAN: AE460260001015862843301

      responses:
        '200':
          description: IBAN validated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                responseCode: "200"
                responseMessage: "OK"
                response:
                  bank_data:
                    countryISO3: ARE
                    bic: ADCBAEAAXXX
                    branch: null
                    bank: ABU DHABI COMMERCIAL BANK
                    address: ADCB BUILDING, SALAM STREET
                    city: ABU DHABI
                    state: null
                    zip: "939"
                    phone: null
                    fax: null
                    www: null
                    email: null
                    country: United Arab Emirates
                    country_iso: AE
                    account: "0012552369920001"
                    bank_code: "003"
                    branch_code: ""
                  validations:
                    chars:
                      code: "006"
                      message: IBAN does not contain illegal characters
                    account:
                      code: "004"
                      message: Account Number check digit is not performed for this bank or branch
                    iban:
                      code: "001"
                      message: IBAN Check digit is correct
                    structure:
                      code: "005"
                      message: IBAN structure is correct
                    length:
                      code: "003"
                      message: IBAN Length is correct
                    country_support:
                      code: "007"
                      message: Country supports IBAN standard

        '401':
          description: Unauthorized - Missing or invalid JWT bearer token

        '400':
          description: Invalid request

        '500':
          description: Internal server error
  /veriphone/verify:
    post:
      tags:
        - VeriPhone
      summary: Verify phone number
      description: Validates a phone number and returns phone metadata.
      operationId: verifyPhone

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneRequest'
            example:
              phone: "971502540238"
      responses:
        '200':
          description: Phone verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                responseCode: "200"
                responseMessage: "OK"
                response:
                  status: success
                  phone: "971502540238"
                  phone_valid: true
                  phone_type: mobile
                  phone_region: United Arab Emirates
                  country: United Arab Emirates
                  country_code: AE
                  country_prefix: "971"
                  international_number: "971502540238"
                  local_number: "0502540238"
                  e164: "971502540238"
                  carrier: du

        '401':
          description: Unauthorized - Missing or invalid JWT bearer token

        '400':
          description: Invalid request

        '500':
          description: Internal server error
  /existing-data/validation:
    post:
      tags:
        - ExistingData
      summary: Validate existing data
      description: Checks if an email, passport, NIN, username, EID, third-party UUID, or mobile number exists in the external system.
      operationId: validateExistingData

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
            examples:
              email:
                summary: Email exists validation
                value:
                  validationType: "EMAIL_EXISTS"
                  referenceNo: "1"
                  email: "solattupuram@alramz.ae"
              passport:
                summary: Passport exists validation
                value:
                  validationType: "PASSPORT_EXISTS"
                  referenceNo: "100"
                  passportNumber: "XY1152302"
              nin:
                summary: NIN exists validation
                value:
                  validationType: "NIN_EXISTS"
                  referenceNo: "200"
                  nin: "1234567890"
              username:
                summary: Username exists validation
                value:
                  validationType: "USERNAME_EXISTS"
                  referenceNo: "300"
                  username: "yasser1"
              eid:
                summary: EID exists validation
                value:
                  validationType: "EID_EXISTS"
                  referenceNo: "400"
                  eidNumber: "784-1989-0751424-8"
              tp_uuid:
                summary: Third-party UUID exists validation
                value:
                  validationType: "TP_UUID_EXISTS"
                  referenceNo: "500"
                  uuid: "101296594260088743321545352540"
                  thirdParty: "Google"
              mobile:
                summary: Mobile number exists validation
                value:
                  validationType: "MOBILE_EXISTS"
                  referenceNo: "600"
                  mobileNumber: "971502573881"

      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                responseCode: "200"
                responseMessage: "OK"
                response:
                  validationType: "EMAIL_EXISTS"
                  exists: true
                  valid: true
                  message: "Email already exists"
                  referenceNo: "1"

        '401':
          description: Unauthorized - Missing or invalid JWT bearer token

        '400':
          description: Invalid request

        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication. Include token in Authorization header.
  schemas:
    IBANRequest:
      type: object
      required:
        - IBAN
      properties:
        IBAN:
          type: string
          description: International Bank Account Number
          minLength: 1
          maxLength: 30
          example: AE460260001015862843301
    GenericResponse:
      type: object
      required:
        - responseCode
        - responseMessage
        - response
      properties:
        responseCode:
          type: string
          example: "200"

        responseMessage:
          type: string
          example: OK

        response:
          type: object

        correlationId:
          type: string
          format: uuid
          description: Correlation ID for request tracing
          example: "550e8400-e29b-41d4-a716-446655440000"
...
    PhoneRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          minLength: 3
          maxLength: 15
          description: Phone number in international format without spaces.
          example: "971502540238"
...
    ValidationRequest:
      type: object
      required:
        - validationType
        - referenceNo
      properties:
        validationType:
          type: string
          enum: [EMAIL_EXISTS, PASSPORT_EXISTS, NIN_EXISTS, USERNAME_EXISTS, EID_EXISTS, TP_UUID_EXISTS, MOBILE_EXISTS]
          description: Type of validation to perform
        referenceNo:
          type: string
          description: Client reference number
        email:
          type: string
          format: email
          description: Email address to validate (required for EMAIL_EXISTS)
        passportNumber:
          type: string
          description: Passport number to validate (required for PASSPORT_EXISTS)
        emiratesId:
          type: string
          description: Emirates ID to validate (reserved for future)
        nin:
          type: string
          description: National ID to validate (required for NIN_EXISTS)
        username:
          type: string
          description: Username to validate (required for USERNAME_EXISTS)
        eidNumber:
          type: string
          description: Emirates ID number to validate (required for EID_EXISTS)
        uuid:
          type: string
          description: Third-party UUID to validate (required for TP_UUID_EXISTS)
        thirdParty:
          type: string
          description: Third-party provider name (required for TP_UUID_EXISTS)
        mobileNumber:
          type: string
          description: Mobile number to validate (required for MOBILE_EXISTS)
...